df: split platform-specific code into platform/{unix,windows}.rs
